refactor: extract useEnableMissingNetwork hook
