+++ ui/pages/bridge/prepare/prepare-bridge-page.test.tsx
@@ -1,20 +1,14 @@
 import React from 'react';
 import { act } from '@testing-library/react';
 import * as reactRouterUtils from 'react-router-dom-v5-compat';
-import * as ReactReduxModule from 'react-redux';
 import { userEvent } from '@testing-library/user-event';
 import { toEvmCaipChainId } from '@metamask/multichain-network-controller';
-import { renderHook } from '@testing-library/react-hooks';
 import { renderWithProvider } from '../../../../test/lib/render-helpers-navigate';
 import configureStore from '../../../store/store';
 import { createBridgeMockStore } from '../../../../test/data/bridge/mock-bridge-store';
 import { CHAIN_IDS } from '../../../../shared/constants/network';
 import { createTestProviderTools } from '../../../../test/stub/provider';
-import * as SelectorsModule from '../../../selectors/multichain/networks';
-import * as ActionsModule from '../../../store/actions';
-import PrepareBridgePage, {
-  useEnableMissingNetwork,
-} from './prepare-bridge-page';
+import PrepareBridgePage from './prepare-bridge-page';
 
 // Mock the bridge hooks
 jest.mock('../hooks/useGasIncluded7702', () => ({
@@ -295,67 +289,5 @@ describe('PrepareBridgePage', () => {
       await userEvent.paste('2abc.131.123456123456123456123456');
     });
     expect(input).toHaveDisplayValue('2.131');
-  });
-});
-
-describe('useEnableMissingNetwork', () => {
-  const arrangeReactReduxMocks = () => {
-    jest
-      .spyOn(ReactReduxModule, 'useSelector')
-      .mockImplementation((selector) => selector({}));
-    jest.spyOn(ReactReduxModule, 'useDispatch').mockReturnValue(jest.fn());
-  };
-
-  const arrange = () => {
-    arrangeReactReduxMocks();
-
-    const mockGetEnabledNetworksByNamespace = jest
-      .spyOn(SelectorsModule, 'getEnabledNetworksByNamespace')
-      .mockReturnValue({
-        '0x1': true,
-        '0xe708': true,
-      });
-    const mockEnableAllPopularNetworks = jest.spyOn(
-      ActionsModule,
-      'setEnabledAllPopularNetworks',
-    );
-
-    return {
-      mockGetEnabledNetworksByNamespace,
-      mockEnableAllPopularNetworks,
-    };
-  };
-
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-
-  it('enables popular network when not already enabled', () => {
-    const mocks = arrange();
-    mocks.mockGetEnabledNetworksByNamespace.mockReturnValue({ '0xe708': true }); // Missing 0x1.
-    const hook = renderHook(() => useEnableMissingNetwork());
-
-    // Act - enable 0x1
-    hook.result.current('0x1');
-
-    // Assert - Adds 0x1 to enabled networks
-    expect(mocks.mockEnableAllPopularNetworks).toHaveBeenCalledWith();
-  });
-
-  it('does not enable popular network if already enabled', () => {
-    const mocks = arrange();
-    const hook = renderHook(() => useEnableMissingNetwork());
-
-    // Act - enable 0x1 (already enabled)
-    hook.result.current('0x1');
-    expect(mocks.mockEnableAllPopularNetworks).not.toHaveBeenCalled();
-  });
-
-  it('does not enable non-popular network', () => {
-    const mocks = arrange();
-    const hook = renderHook(() => useEnableMissingNetwork());
-
-    hook.result.current('0x1111'); // not popular network
-    expect(mocks.mockEnableAllPopularNetworks).not.toHaveBeenCalled();
   });
 });
